refactor: constants 폴더 생성
